added button to start and stop recording
right now the code has a button to start and stop recording. when the recording is stopped it automatically downloads the video to the users pc. instead of this we want to upload it to the DB. and instead of the button we want the recording to start AFTER the "models loaded" debug and stop when the exam stops automatiucally.